Reformat/Refactor SearchView.
